Use libponyc's package_init for discovering stdlib path in pony-lsp
This works under the assumption that the pony-lsp binary is placed right next to the ponyc binary in releases.
This might not reliably work in test binaries, like when building and running the tools tests, it depends on whether we have a stdlib at a known path on the system and the test binary location. I didn't yet find a good way to separate this out for tests.